test(cli): extract unit move preview play tests
Move unit-move-preview command coverage out of the play monolith into the foldered play command test suite, preserving the neutral relationship-policy assertions and using a narrow local movement preview fake. Remove the now-dead unit-move-preview fake handler and fixture from the monolith so ownership is not duplicated without an adjacent consumer.

Proof boundary: test-only modularization. This layer changes no runtime direct-control behavior, play-thread flow, broad fake tuner ownership, oRPC transport, relationship authority policy, or broader play-suite folder organization.
